% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate_trgi_data}
\alias{simulate_trgi_data}
\title{Simulate mixed-type data with global and local nonlinearity}
\usage{
simulate_trgi_data(n = 400, seed = 42)
}
\arguments{
\item{n}{Sample size}

\item{seed}{Random seed}
}
\value{
A list with elements:
\describe{
\item{X}{data.frame of predictors}
\item{y}{numeric response}
}
}
\description{
Simulate mixed-type data with global and local nonlinearity
}
