% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trgi_fit.R
\name{trgi_fit}
\alias{trgi_fit}
\title{Fit TRGI-MRA++ model}
\usage{
trgi_fit(
  X,
  y,
  cv_folds = 5,
  micro_candidates = c("rf", "gam", "xgb"),
  seed = 123
)
}
\arguments{
\item{X}{data.frame of predictors (mixed numeric/categorical)}

\item{y}{numeric response}

\item{cv_folds}{number of CV folds}

\item{micro_candidates}{micro learners to consider}

\item{seed}{random seed}
}
\value{
object of class 'trgi_mra'
}
\description{
Fit TRGI-MRA++ model
}
